# src/AppForSEII2526.Web/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AppForMovies.API",
    "description": "This API provides services for renting and purchasing movies",
    "contact": {
      "name": "Software Engineering II Team",
      "email": "isii@on.uclm.es"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/license/mit/"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Maintenance/GetMaintenances": {
      "get": {
        "tags": [
          "Maintenance"
        ],
        "operationId": "GetMaintenances",
        "parameters": [
          {
            "name": "nombre",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tipo",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MantenimientoDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MantenimientoDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MantenimientoDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MantenimientoDTO": {
        "required": [
          "nombre",
          "numeroDias",
          "precio",
          "tipo"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "minLength": 1,
            "type": "string"
          },
          "tipo": {
            "minLength": 1,
            "type": "string"
          },
          "precio": {
            "maximum": 500,
            "minimum": 1,
            "type": "number",
            "format": "currency"
          },
          "numeroDias": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}